# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
  /api/v1/tariff-catalog/{id}:
    get:
      tags:
      - Management-data-controller
      operationId: getTariff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TariffResponse"
    put:
      tags:
      - Management-data-controller
      operationId: updateTariff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TariffUpsertRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TariffResponse"
    delete:
      tags:
      - Management-data-controller
      operationId: deleteTariff
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/tariff-options/{id}:
    get:
      tags:
      - Management-data-controller
      operationId: getTariffOption
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TariffOptionResponse"
    put:
      tags:
      - Management-data-controller
      operationId: updateTariffOption
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TariffOptionUpsertRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TariffOptionResponse"
    delete:
      tags:
      - Management-data-controller
      operationId: deleteTariffOption
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
  /api/v1/tariff-categories/{id}:
    get:
      tags:
      - Management-data-controller
      operationId: getTariffCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IdNameResponse"
    put:
      tags:
      - Management-data-controller
      operationId: updateTariffCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NameRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IdNameResponse"
    delete:
      tags:
      - Management-data-controller
      operationId: deleteTariffCategory
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
...
  /api/v1/subscribers/{subscriberId}/tariff/change:
    post:
      tags:
      - tariff-controller
      operationId: changeTariff
      parameters:
      - name: subscriberId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeTariffRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ChangeTariffResponse"
...
  /api/v1/tariff-catalog:
    get:
      tags:
      - Management-data-controller
      operationId: getTariffs
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TariffResponse"
    post:
      tags:
      - Management-data-controller
      operationId: createTariff
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TariffUpsertRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TariffResponse"
  /api/v1/tariff-options:
    get:
      tags:
      - Management-data-controller
      operationId: getTariffOptions
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TariffOptionResponse"
    post:
      tags:
      - Management-data-controller
      operationId: createTariffOption
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TariffOptionUpsertRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TariffOptionResponse"
  /api/v1/tariff-categories:
    get:
      tags:
      - Management-data-controller
      operationId: getTariffCategories
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IdNameResponse"
    post:
      tags:
      - Management-data-controller
      operationId: createTariffCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NameRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IdNameResponse"
...
  /api/v1/tariffs:
    get:
      tags:
      - tariff-controller
      operationId: findTariffs
      parameters:
      - name: categoryId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: query
        in: query
        required: false
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TariffSummaryDto"
  /api/v1/tariffs/{tariffId}:
    get:
      tags:
      - tariff-controller
      operationId: getTariffDetails
      parameters:
      - name: tariffId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TariffDetailsResponse"
...
  /api/v1/subscribers/{subscriberId}/tariff:
    get:
      tags:
      - tariff-controller
      operationId: getSubscriberTariffInfo
      parameters:
      - name: subscriberId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TariffInfoResponse"
...
components:
  schemas:
    TariffUpsertRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        monthlyFee:
          type: number
          minimum: 0.00
        switchFee:
          type: number
          minimum: 0.00
        customizable:
          type: boolean
        pdfUrl:
          type: string
        categoryId:
          type: integer
          format: int64
      required:
      - categoryId
      - description
      - monthlyFee
      - name
      - pdfUrl
      - switchFee
    TariffResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        monthlyFee:
          type: number
        switchFee:
          type: number
        customizable:
          type: boolean
        pdfUrl:
          type: string
        categoryId:
          type: integer
          format: int64
    TariffOptionUpsertRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
          minimum: 0.00
        tariffId:
          type: integer
          format: int64
      required:
      - description
      - name
      - price
      - tariffId
    TariffOptionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
        tariffId:
          type: integer
          format: int64
    NameRequest:
      type: object
      properties:
        name:
          type: string
      required:
      - name
    IdNameResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
...
    ChangeTariffRequest:
      type: object
      properties:
        targetTariffId:
          type: integer
          format: int64
        options:
          type: object
          additionalProperties:
            type: string
      required:
      - targetTariffId
    BillingTransactionDto:
      type: object
      properties:
        type:
          type: string
        amount:
          type: number
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    ChangeTariffResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        previousTariff:
          $ref: "#/components/schemas/TariffSummaryDto"
        newTariff:
          $ref: "#/components/schemas/TariffSummaryDto"
        selectedOptions:
          type: object
          additionalProperties:
            type: string
        balance:
          type: number
        billingTransactions:
          type: array
          items:
            $ref: "#/components/schemas/BillingTransactionDto"
        notification:
          $ref: "#/components/schemas/NotificationDto"
    NotificationDto:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
        success:
          type: boolean
        createdAt:
          type: string
          format: date-time
    TariffSummaryDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        category:
          type: string
        monthlyFee:
          type: number
...
    TariffDetailsResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        category:
          type: string
        monthlyFee:
          type: number
        switchFee:
          type: number
        customizable:
          type: boolean
        pdfUrl:
          type: string
        options:
          type: array
          items:
            $ref: "#/components/schemas/TariffOptionDto"
    TariffOptionDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        price:
          type: number
...
    TariffInfoResponse:
      type: object
      properties:
        subscriberId:
          type: integer
          format: int64
        subscriberPhone:
          type: string
        subscriberName:
          type: string
        balance:
          type: number
        currentTariff:
          $ref: "#/components/schemas/TariffSummaryDto"
...